MAINT: Remove MongoConfig.port.
It's possible to have a cluster where each host is on a different port,
which makes having a single integer representing the port a dumb way to
represent the connection info.

We now require ports to be specified as part of the host string, in the
form {hostname}:{port}.

An alternative to this would be to allow/require passing a list of ports
instead of a single integer, but at that point you're separating
information into two lists that you probably would rather have bundled
together.